Changes for readr 2.0.0
In readr 2.0.0 the way that literal string data is passed to readr has
changed. Previously any length 2+ character vector was treated as
literal data, now these vectors are treated as multiple file names to be
read into one dataset.

To specify literal data you can use wrap it in `I()`, which is what this
change does.